Refactor configurations, test, and constants
